pyproject.toml
[project]
name = "blackhole-io"
version = "0.1.2"
description = "Universal file storage interface for Python"
readme = "README.md"
authors = [
    { name = "fobiasmog", email = "zbrejkin@yandex.ru" }
]
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "boto3>=1.40.11",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0",
    "starlette>=0.47.2",
]

[dependency-groups]
dev = [
    "isort>=8.0.0",
    "ruff>=0.15.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/fobiasmog/blackhole"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
